requirements.txt
fastapi>=0.115,<0.116
uvicorn>=0.30,<0.31
SQLAlchemy>=2.0,<2.1
psycopg[binary]>=3.1,<3.2
alembic>=1.13,<1.14
python-dotenv>=1.0,<1.1
pytest>=8.0,<9.0
httpx>=0.27,<0.28
